+++ src/main/java/org/ehrbase/fhirbridge/ehr/converter/specific/geccoVirologischerbefund/PCRCompositionConverter.java
@@ -1,4 +1,4 @@
-package org.ehrbase.fhirbridge.ehr.converter.specific.virologischerBefund;
+package org.ehrbase.fhirbridge.ehr.converter.specific.geccoVirologischerbefund;
 
 import ca.uhn.fhir.rest.server.exceptions.UnprocessableEntityException;
 import org.ehrbase.fhirbridge.ehr.converter.generic.ObservationToCompositionConverter;
